* release/v2.10: (191 commits)
  Increment the version number
  Restructure tagged trait tables
  Show archived taggedtraits to staff on tag+study page
  Fix columns in TaggedTraitTableWithDCCReviewButton
  Improve tests of TaggedTraitDetail
  Clean up test names and docstrings
  Show archived status on TaggedTraitDetail
  Add tests of detail links on TaggedTraitByTagAndStudyList
  Move detail links column to restricted tables
  Make TaggedTraitDetail restricted
  Clean up permissions mixins in tags.views
  Auto-archive after agree response
  Handle archiving in StudyResponseCheckMixin
  Display archived status on DCCReviewNeedFollowupList
  Hide response buttons for archived tagged traits
  Add tests for archiving in DCCReviewNeedFollowupList
  Fix queries in DCCReviewNeedFollowupCounts
  Account for archiving in DCCReviewNeedFollowupCounts
  Change display of other tags in taggedtrait_info
  Remove use of comment argument for DCCReviewFactory
  ...

+++ tags/models.py
@@ -1,5 +1,6 @@
 """Models for the tags app."""
 
+from django.apps import apps
 from django.conf import settings
 from django.db import models
 from django.urls import reverse
@@ -19,7 +20,7 @@ class Tag(TimeStampedModel):
     description = models.TextField()
     instructions = models.TextField()
     creator = models.ForeignKey(settings.AUTH_USER_MODEL, blank=True, on_delete=models.CASCADE)
-    traits = models.ManyToManyField('trait_browser.SourceTrait', through='TaggedTrait')
+    all_traits = models.ManyToManyField('trait_browser.SourceTrait', through='TaggedTrait', related_name='all_tags')
 
     class Meta:
         verbose_name = 'phenotype tag'
@@ -43,13 +44,28 @@ class Tag(TimeStampedModel):
     def get_absolute_url(self):
         return reverse('tags:tag:detail', args=[self.pk])
 
+    @property
+    def archived_traits(self):
+        """Return queryset of archived traits tagged with this tag."""
+        archived_tagged_traits = apps.get_model('tags', 'TaggedTrait').objects.archived().filter(tag=self)
+        return apps.get_model('trait_browser', 'SourceTrait').objects.filter(
+            pk__in=archived_tagged_traits.values_list('trait__pk', flat=True))
+
+    @property
+    def non_archived_traits(self):
+        """Return queryset of non-archived traits tagged with this tag."""
+        non_archived_tagged_traits = apps.get_model('tags', 'TaggedTrait').objects.non_archived().filter(tag=self)
+        return apps.get_model('trait_browser', 'SourceTrait').objects.filter(
+            pk__in=non_archived_tagged_traits.values_list('trait__pk', flat=True))
+
 
 class TaggedTrait(TimeStampedModel):
     """Intermediate model for connecting Tags and SourceTraits."""
 
-    trait = models.ForeignKey('trait_browser.SourceTrait', on_delete=models.CASCADE)
-    tag = models.ForeignKey(Tag, on_delete=models.CASCADE)
+    trait = models.ForeignKey('trait_browser.SourceTrait', on_delete=models.CASCADE, related_name="all_taggedtraits")
+    tag = models.ForeignKey(Tag, on_delete=models.CASCADE, related_name="all_taggedtraits")
     creator = models.ForeignKey(settings.AUTH_USER_MODEL, blank=True, on_delete=models.CASCADE)
+    archived = models.BooleanField(default=False)
 
     # Managers/custom querysets.
     objects = querysets.TaggedTraitQuerySet.as_manager()
@@ -60,16 +76,40 @@ class TaggedTrait(TimeStampedModel):
 
     def __str__(self):
         """Pretty printing."""
-        return 'variable {} tagged {}'.format(self.trait.i_trait_name, self.tag.lower_title)
+        return 'variable {} tagged {}'.format(self.trait.i_trait_name, self.tag.title)
 
     def get_absolute_url(self):
         return reverse('tags:tagged-traits:pk:detail', args=[self.pk])
 
+    def archive(self):
+        """Set the TaggedTrait to archived.
+
+        Other than tests, this method should only be used on tagged traits that have
+        linked DCCReviews. Using it on unreviewed tagged traits makes some error
+        messages nonsensical.
+        """
+        self.archived = True
+        self.save()
+
+    def unarchive(self):
+        """Set the TaggedTrait to unarchived.
+
+        Other than tests, this method should only be used on tagged traits that have
+        linked DCCReviews. Using it on unreviewed tagged traits makes some error
+        messages nonsensical.
+        """
+        self.archived = False
+        self.save()
+
     def delete(self, *args, **kwargs):
         """Only allow unreviewed TaggedTrait objects to be deleted."""
         if hasattr(self, 'dcc_review'):
-            raise DeleteNotAllowedError("Cannot delete a reviewed TaggedTrait.")
-        super().delete(*args, **kwargs)
+            if self.dcc_review.status == self.dcc_review.STATUS_CONFIRMED:
+                raise DeleteNotAllowedError("Cannot delete a reviewed TaggedTrait.")
+            else:
+                self.archive()
+        else:
+            self.hard_delete()
 
     def hard_delete(self, *args, **kwargs):
         """Delete objects that cannot be deleted with overriden delete method."""
@@ -90,8 +130,42 @@ class DCCReview(TimeStampedModel):
     comment = models.TextField(blank=True)
     creator = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
 
+    # Managers/custom querysets.
+    objects = querysets.DCCReviewQuerySet.as_manager()
+
     class Meta:
         verbose_name = 'dcc review'
 
+    def __str__(self):
+        return 'Review of {}'.format(self.tagged_trait)
+
     def get_absolute_url(self):
         return self.tagged_trait.get_absolute_url()
+
+    def delete(self, *args, **kwargs):
+        """Only allow DCCReview objects without a StudyResponse to be deleted."""
+        if hasattr(self, 'study_response'):
+            raise DeleteNotAllowedError("Cannot delete a DCCReview with a study response.")
+        super().delete(*args, **kwargs)
+
+    def hard_delete(self, *args, **kwargs):
+        """Delete objects that cannot be deleted with overriden delete method."""
+        super().delete(*args, **kwargs)
+
+
+class StudyResponse(TimeStampedModel):
+    """Model to allow study users to respond to a DCCReview."""
+
+    dcc_review = models.OneToOneField(DCCReview, on_delete=models.CASCADE, related_name='study_response')
+    STATUS_DISAGREE = 0
+    STATUS_AGREE = 1
+    STATUS_CHOICES = (
+        (STATUS_AGREE, 'Agree'),
+        (STATUS_DISAGREE, 'Disagree'),
+    )
+    status = models.IntegerField(choices=STATUS_CHOICES)
+    comment = models.TextField(blank=True)
+    creator = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
+
+    class Meta:
+        verbose_name = 'study response'
